8259723: Move Table class to formats.html package
